requirements.txt
backports.tarfile
certifi
charset-normalizer
exceptiongroup
ffmpeg
flatbuffers
gast
idna
importlib-metadata
ipykernel
jaraco.collections
libclang
matplotlib
mdurl
namex
opencv-python
opt-einsum
optree
pandas
pretty-errors
pyinputplus
python-mnist
scikit-learn
tensorboard
termcolor
tomli
torchaudio
torchvision
tqdm
urllib3
websockets
wrapt
